Put CAD service tests in s seperate class
