+++ JplApiTesting/JplApiTesting/ApiObjectModels/CAD/Services/CadSpecificClassService.cs
@@ -7,7 +7,7 @@ namespace JplApiTesting.ApiObjectModels.CAD.Services
         public CadSpecificClassService(string specificClass)
         {
             if (specificClass == string.Empty)
-                throw new ArgumentException("The body cannot be an empty string");
+                throw new ArgumentException("The class cannot be an empty string");
 
             liveCurrent = callManager.GetSpecificClassData(specificClass);
