+++ JplApiTesting/JplApiTesting/JplApiTesting.csproj
@@ -77,6 +77,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tests\CAD\CadInvalidBodyTypeTests.cs" />
     <Compile Include="Tests\CAD\CadInvalidParamTests.cs" />
+    <Compile Include="Tests\CAD\CadServiceTests.cs" />
     <Compile Include="Tests\CAD\CadSpecificBodyTests.cs" />
     <Compile Include="Tests\CAD\CadLimitedRequestTests.cs" />
     <Compile Include="Tests\CAD\CadSpecificClassTests.cs" />
